Support CustomData
As of the latest fabric update a new element called "CustomData" can be
defined for the Windows/Linux Provisioning Configuration data:
http://msdn.microsoft.com/en-us/library/windowsazure/jj157194.aspx

This patch adds a new parameter called "--custom-data" that points to a
file.  The contents of the file will be base64 encoded and added to the
Windows/Linux Provisioning Configuration data.