Can use function for secure env. variable
When recording, we can now set a function as the value for the
'secure' field in the requiredEnvironments object. If a function
is used, this function will be called when recording to replace
the value of the environment variable. If it's any other truthy
value the value recorded will be 'dummy' as before.