Made env add, env set commands data driven.
The options for account env add and set are now table driven
rather than hard-coded in the command options. This way going
forward we don't have to worry about changing these commands
if we add or remove options on environments.

This is a breaking change on the options names for the commands,
they now line up with the actual parameter names on the environment
object.