Fixed site log command, rerecorded tests.
The site log set command wasn't pulling the subscription ID correctly,
resulting in an extra cache file that can get out of sync with the real
run. This commit fixes the command, adds error checking to the caching
logic to detect this failure in the future, and rerecords all the site
mocks since they were potentially affected by this bug.