BUG:2037247 - [NodeJS] GetShareByPrefixUsingWildCardTest failed because * is taken as valid character by NodeJS SDK
Both NodeJS SDK and .Net xSCL didn't validate share prefix for their purpose. Their design seems to just return nothing if a prefix is not valid (and could never be a prefix of any possible share name).

For our cmdlet, I believe it would be better to fail earlier. So I did this change to validate the prefix on our own and follows the same behavior of powershell cmdlets.