vm create: check virtual network against existing deployment
When the user specifies the --connect option to the vm create command,
and there is an existing deployment, the new VM will use whatever
virtual network the existing deployment belongs to.
If the user also passes --virtual-network-name, we need to check
that the given virtual network matches the existing deployment's.

I also made --virtual-network-name mandatory when --subnet-names is given.

This fixes https://github.com/MSOpenTech/azure-sdk-tools-xplat/issues/65